final cart functionality
